pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "storytime"
version = "0.1.0"
description = "Audiobook parsing and TTS generation toolkit"
authors = [{ name = "Storytime" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pydantic>=2.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.82.0",
    "python-dotenv>=1.0.0",
    "pydub>=0.25.1",
    "requests>=2.31.0",
    "pytest>=7.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic[email]",
    "pydantic-settings>=2.0.0",
    "elevenlabs>=2.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "junjo>=0.45.0",
    "opentelemetry-sdk>=1.33.0",
    "opentelemetry-exporter-otlp>=1.33.0",
    "opentelemetry-exporter-otlp-proto-http>=1.33.0",
    "celery[redis]",
    "psycopg2-binary",
    "asyncpg",
    "sqlalchemy",
    "boto3>=1.34.0",
    "greenlet (>=3.2.2,<4.0.0)",
    "PyJWT[crypto]>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
development = [
    "pytest",
    "ruff>=0.11.0",
    "mypy",
    "pre-commit",
    "alembic>=1.13.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.1"
pytest-asyncio = "^1.0.0"

